Fixes from GitHub Review
https://github.com/aerogear/aerogear-android-cookbook/pull/30
