+++ AuthExamples/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="org.jboss.aerogear.authexamples" >
+    package="org.jboss.aerogear.android.cookbook.authexamples" >
 
     <uses-permission android:name="android.permission.INTERNET"/>
 
@@ -18,8 +18,10 @@
         <activity
             android:name=".MainActivity"
             android:label="@string/title_activity_main" >
-            <action android:name="android.intent.action.MAIN" />
-            <category android:name="android.intent.category.LAUNCHER" />
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
         </activity>
     </application>
