+++ AuthExamples/app/src/main/java/org/jboss/aerogear/android/cookbook/authexamples/MainActivity.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.aerogear.authexamples;
+package org.jboss.aerogear.android.cookbook.authexamples;
 
 import android.app.Activity;
 import android.content.Intent;
