+++ AuthExamples/app/src/main/res/layout/authentication.xml
