+++ AuthExamples/app/src/main/res/layout/authentication_with_data.xml
@@ -63,3 +63,4 @@
     </LinearLayout>
 
 </LinearLayout>
+
